requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
black==23.3.0
click==8.1.3
fastapi==0.95.1
h11==0.14.0
hpack==4.0.0
httptools==0.5.0
hyperframe==6.0.1
idna==3.4
iso8601==1.1.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pydantic==1.10.7
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
